requirements.txt
langchain-community
jq
langchain-openai
pinecone
streamlit
python-dotenv
openai
sentence-transformers
langchain-huggingface